+++ src/connection/impls/file.rs
@@ -1,3 +1,4 @@
+use std::fmt::{Display, Formatter};
 use std::{
     ffi::c_int,
     io,
@@ -6,7 +7,7 @@ use std::{
 };
 
 use crate::{
-    connection::{IpmiConnection, Message, Request, Response},
+    connection::{Address, IpmiConnection, Message, Request, RequestTargetAddress, Response},
     NetFn,
 };
 
@@ -105,6 +106,58 @@ mod ioctl {
 
     ioctl_readwrite!(ipmi_recv_msg_trunc, IPMI_IOC_MAGIC, 11, IpmiRecv);
     ioctl_read!(ipmi_send_request, IPMI_IOC_MAGIC, 13, IpmiRequest);
+    ioctl_read!(ipmi_get_my_address, IPMI_IOC_MAGIC, 18, u32);
+}
+
+#[repr(C)]
+enum IpmiAddr {
+    SysIface(IpmiSysIfaceAddr),
+    Ipmb(IpmiIpmbAddr),
+}
+
+impl IpmiAddr {
+    fn ptr(&mut self) -> *mut u8 {
+        match self {
+            IpmiAddr::SysIface(ref mut bmc_addr) => std::ptr::addr_of_mut!(*bmc_addr) as *mut u8,
+            IpmiAddr::Ipmb(ref mut ipmb_addr) => std::ptr::addr_of_mut!(*ipmb_addr) as *mut u8,
+        }
+    }
+    fn size(&self) -> u32 {
+        match self {
+            IpmiAddr::SysIface(_) => core::mem::size_of::<IpmiSysIfaceAddr>() as u32,
+            IpmiAddr::Ipmb(_) => core::mem::size_of::<IpmiIpmbAddr>() as u32,
+        }
+    }
+}
+
+impl From<RequestTargetAddress> for IpmiAddr {
+    fn from(value: RequestTargetAddress) -> Self {
+        match value {
+            RequestTargetAddress::Bmc(lun) => {
+                IpmiAddr::SysIface(IpmiSysIfaceAddr::bmc(lun.value()))
+            }
+            RequestTargetAddress::BmcOrIpmb(addr, channel, lun) => {
+                IpmiAddr::Ipmb(IpmiIpmbAddr::new(channel.0 as i16, addr.0, lun.value()))
+            }
+        }
+    }
+}
+
+impl Display for IpmiAddr {
+    fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
+        match self {
+            IpmiAddr::SysIface(addr) => {
+                write!(f, "System interface (LUN: {})", addr.lun)
+            }
+            IpmiAddr::Ipmb(addr) => {
+                write!(
+                    f,
+                    "IPMB target (Channel: {}, Target: {}, LUN: {})",
+                    addr.channel, addr.target_addr, addr.lun
+                )
+            }
+        }
+    }
 }
 
 #[repr(C)]
@@ -113,6 +166,28 @@ pub struct IpmiSysIfaceAddr {
     ty: i32,
     channel: i16,
     lun: u8,
+}
+
+#[repr(C)]
+#[derive(Clone, Debug, PartialEq)]
+pub struct IpmiIpmbAddr {
+    ty: i32,
+    channel: i16,
+    target_addr: u8,
+    lun: u8,
+}
+
+impl IpmiIpmbAddr {
+    const IPMI_IPMB_ADDR_TYPE: i32 = 0x01;
+
+    pub const fn new(channel: i16, target_addr: u8, lun: u8) -> Self {
+        Self {
+            ty: Self::IPMI_IPMB_ADDR_TYPE,
+            channel,
+            target_addr,
+            lun,
+        }
+    }
 }
 
 impl IpmiSysIfaceAddr {
@@ -132,6 +207,7 @@ pub struct File {
     inner: std::fs::File,
     recv_timeout: Duration,
     seq: i64,
+    my_addr: Address,
 }
 
 impl File {
@@ -140,13 +216,36 @@ impl File {
     }
 
     pub fn new(path: impl AsRef<std::path::Path>, recv_timeout: Duration) -> io::Result<Self> {
+        let mut inner = std::fs::File::open(path)?;
+
+        let my_addr = match Self::load_my_address_from_file(&mut inner) {
+            Ok(addr) => addr,
+            Err(e) => {
+                log::warn!("Failed to get local address, defaulting to 0x20: {:?}", e);
+                Address(0x20)
+            }
+        };
         let me = Ok(Self {
-            inner: std::fs::File::open(path)?,
+            inner,
             recv_timeout,
             seq: 0,
+            my_addr,
         });
 
         me
+    }
+
+    fn load_my_address_from_file(file: &mut std::fs::File) -> io::Result<Address> {
+        let mut my_addr: u32 = 8;
+        unsafe { ioctl::ipmi_get_my_address(file.as_raw_fd(), std::ptr::addr_of_mut!(my_addr))? };
+        if let Ok(addr) = u8::try_from(my_addr) {
+            Ok(Address(addr))
+        } else {
+            Err(io::Error::new(
+                io::ErrorKind::Other,
+                format!("ipmi_get_my_address returned non-u8 address: {}", my_addr),
+            ))
+        }
     }
 }
 
@@ -156,7 +255,13 @@ impl IpmiConnection for File {
     type Error = io::Error;
 
     fn send(&mut self, request: &mut Request) -> io::Result<()> {
-        let mut bmc_addr = IpmiSysIfaceAddr::bmc(request.lun().value());
+        let mut addr: IpmiAddr = match request.target() {
+            RequestTargetAddress::BmcOrIpmb(a, _, lun) if a == self.my_addr => {
+                RequestTargetAddress::Bmc(lun)
+            }
+            x => x,
+        }
+        .into();
 
         let netfn = request.netfn_raw();
         let cmd = request.cmd();
@@ -166,19 +271,20 @@ impl IpmiConnection for File {
         let data_len = data.len() as u16;
         let ptr = data.as_mut_ptr();
 
+        log::debug!("Sending request (netfn: 0x{netfn:02X}, cmd: 0x{cmd:02X}) to {addr}");
+        let ipmi_message = IpmiMessage {
+            netfn,
+            cmd,
+            data_len,
+            data: ptr,
+        };
         let mut request = IpmiRequest {
-            addr: std::ptr::addr_of_mut!(bmc_addr) as *mut u8,
-            addr_len: core::mem::size_of::<IpmiSysIfaceAddr>() as u32,
+            addr: addr.ptr(),
+            addr_len: addr.size(),
             msg_id: seq,
-            message: IpmiMessage {
-                netfn,
-                cmd,
-                data_len,
-                data: ptr,
-            },
+            message: ipmi_message,
         };
 
-        log::debug!("Sending request (netfn: 0x{netfn:02X}, cmd: 0x{cmd:02X})");
         request.log(log::Level::Trace);
 
         // SAFETY: we send a mut pointer to an owned struct (`request`),
@@ -191,7 +297,7 @@ impl IpmiConnection for File {
         #[allow(clippy::drop_non_drop)]
         drop(request);
         #[allow(clippy::drop_non_drop)]
-        drop(bmc_addr);
+        drop(addr);
 
         Ok(())
     }