+++ src/connection/request.rs
@@ -3,14 +3,14 @@ use crate::connection::{LogicalUnit, NetFn};
 use super::Message;
 
 pub struct Request {
-    lun: LogicalUnit,
+    target: RequestTargetAddress,
     message: Message,
 }
 
 impl Request {
-    pub const fn new(request: Message, lun: LogicalUnit) -> Self {
+    pub const fn new(request: Message, target: RequestTargetAddress) -> Self {
         Self {
-            lun,
+            target,
             message: request,
         }
     }
@@ -23,10 +23,6 @@ impl Request {
         self.message.netfn_raw()
     }
 
-    pub fn lun(&self) -> LogicalUnit {
-        self.lun
-    }
-
     pub fn cmd(&self) -> u8 {
         self.message.cmd
     }
@@ -37,5 +33,30 @@ impl Request {
 
     pub fn data_mut(&mut self) -> &mut [u8] {
         self.message.data_mut()
+    }
+
+    pub fn target(&self) -> RequestTargetAddress {
+        self.target
+    }
+}
+
+#[derive(Copy, Clone, Debug, PartialEq)]
+pub struct Address(pub u8);
+#[derive(Copy, Clone, Debug, PartialEq)]
+pub struct Channel(pub u8);
+
+#[derive(Copy, Clone, Debug, PartialEq)]
+pub enum RequestTargetAddress {
+    Bmc(LogicalUnit),
+    BmcOrIpmb(Address, Channel, LogicalUnit),
+}
+
+impl RequestTargetAddress {
+    pub fn lun(&self) -> LogicalUnit {
+        match self {
+            RequestTargetAddress::Bmc(lun) | RequestTargetAddress::BmcOrIpmb(_, _, lun) => {
+                lun.clone()
+            }
+        }
     }
 }