+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Unreleased
-No changes.
+* Support for sending bridged IPMB messages for sensors that are not available on the system
+  interface for File-based connections. `GetSensorReading::new()` and `GetSensorReading::for_sensor()`
+  have been replaced with`GetSensorReading::for_sensor_key()` which now takes a `&Sensorkey`. ([#6])
+
+[#6]: https://github.com/datdenkikniet/ipmi-rs/pull/6
 
 # [0.2.1](https://github.com/datdenkikniet/ipmi-rs/tree/v0.2.1)